pyproject.toml
[project]
name = "mcphost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.51.0",
    "httpx[socks]>=0.28.1",
    "mcp>=1.6.0",
    "openai>=1.76.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
